part-1-intro/solution/requirements.txt
# Strands Agents Workshop - Part 1: Introduction to Agents
# Core dependencies for building your first agent

strands-agents>=0.1.0
python-dotenv>=1.0.0
